refactor: add semi_analytic module
