+++ scripts/plot_first_order.py
@@ -5,10 +5,10 @@ from math import *  # noqa: F401, F403
 import numpy as np
 from matplotlib import colors
 
-from photosurfactant.first_order import FirstOrder, Variables
+from photosurfactant.semi_analytic.first_order import FirstOrder, Variables
 from photosurfactant.fourier import convolution_coeff, fourier_series_coeff
 from photosurfactant.intensity_functions import *  # noqa: F401, F403
-from photosurfactant.leading_order import LeadingOrder
+from photosurfactant.semi_analytic.leading_order import LeadingOrder
 from photosurfactant.parameters import Parameters, PlottingParameters
 from photosurfactant.utils import (
     first_order_parser,
