+++ scripts/plot_spectrum.py
@@ -3,10 +3,10 @@ from argparse import ArgumentDefaultsHelpFormatter, ArgumentParser
 
 import numpy as np
 
-from photosurfactant.first_order import FirstOrder, Variables
+from photosurfactant.semi_analytic.first_order import FirstOrder, Variables
 from photosurfactant.fourier import convolution_coeff, fourier_series_coeff
 from photosurfactant.intensity_functions import mollifier, square_wave
-from photosurfactant.leading_order import LeadingOrder
+from photosurfactant.semi_analytic.leading_order import LeadingOrder
 from photosurfactant.parameters import Parameters, PlottingParameters
 from photosurfactant.utils import (
     first_order_parser,
