+++ src/photosurfactant/__init__.py
@@ -1,3 +1,1 @@
-from .first_order import FirstOrder, Variables
-from .leading_order import LeadingOrder
 from .parameters import Parameters
